スリム化のためcontroller model内のメソッドをconcernへ移動 + 不要なコードの修正
