+++ app/models/user.rb
@@ -1,5 +1,6 @@
 class User < ApplicationRecord
   include AvatarResize
+  include Omniauth
 
   devise :database_authenticatable,
          :registerable,
@@ -30,23 +31,4 @@ class User < ApplicationRecord
             size: {
               less_than: 3.megabytes, message: "のファイルサイズは3MBまでです。",
             }
-
-  class << self
-    def find_for_oauth(auth)
-      user = User.where(uid: auth.uid, provider: auth.provider).first
-      user ||= User.new( # rubocop:disable Lint/UselessAssignment
-        uid: auth.uid,
-        provider: auth.provider,
-        username: auth.info.name,
-        email: User.dummy_email(auth),
-        password: Devise.friendly_token[0, 20]
-      )
-    end
-
-    protected
-
-    def dummy_email(auth)
-      "#{auth.uid}-#{auth.provider}@example.com"
-    end
-  end
 end
