+++ app/views/layouts/_footer.html.slim
@@ -1,2 +1,2 @@
-== render "layouts/footer_navbar"
-== render "layouts/footer_bottombar"
+= render "layouts/footer_navbar"
+= render "layouts/footer_bottombar"
