+++ app/views/layouts/_header.html.slim
@@ -1,2 +1,2 @@
-== render "layouts/header_topbar"
-== render "layouts/header_navbar"
+= render "layouts/header_topbar"
+= render "layouts/header_navbar"
