+++ app/views/layouts/_header_navbar.html.slim
@@ -2,4 +2,4 @@ nav.navbar.navbar-expand-md.bg-white.d-md-block.d-none
   .container
     .collapse.navbar-collapse.justify-content-center
       ul.navbar-nav
-        == render "layouts/header_navbar_list", nav_class: "mr-3", nav_color: nil
+        = render "layouts/header_navbar_list", nav_class: "mr-3", nav_color: nil
