+++ app/views/layouts/_header_topbar.html.slim
@@ -26,7 +26,7 @@ nav.navbar.navbar-expand-md.bg-dark.navbar-dark
             = link_to destroy_user_session_path, method: :delete, class: "nav-link text-white" do
               i.fas.fa-sign-out-alt.mr-1
               | ログアウト
-          == render "layouts/header_navbar_list", nav_class: "d-block d-md-none", nav_color: "text-white"
+          = render "layouts/header_navbar_list", nav_class: "d-block d-md-none", nav_color: "text-white"
 
         - else
           li.nav-item
@@ -37,4 +37,4 @@ nav.navbar.navbar-expand-md.bg-dark.navbar-dark
             = link_to new_user_session_path, class: "nav-link text-white" do
               i.fas.fa-sign-in-alt.mr-2
               | ログイン
-          == render "layouts/header_navbar_list", nav_class: "d-block d-md-none", nav_color: "text-white"
+          = render "layouts/header_navbar_list", nav_class: "d-block d-md-none", nav_color: "text-white"
