+++ app/views/users/confirmations/new.html.slim
@@ -8,4 +8,4 @@
   .card-body
     em アカウント登録したメールアドレスをご入力してください
     = form_with model: resource, url: account_confirmation_path, local: true do |f|
-      == render "users/shared/email_form", f: f
+      = render "users/shared/email_form", f: f
