+++ app/views/users/passwords/new.html.slim
@@ -9,4 +9,4 @@
     em
       | アカウント登録したメールアドレスをご入力してください
     = form_with model: resource, url: reset_password_path, local: true do |f|
-      == render "users/shared/email_form", f: f
+      = render "users/shared/email_form", f: f
