+++ app/views/users/shared/_sidebar.html.slim
