+++ app/views/users/shared/_sns_account.html.slim
@@ -5,14 +5,14 @@
         = "SNSアカウントで#{msg}"
     .card-body.mx-auto
       .btn-sns
-        = link_to user_facebook_omniauth_authorize_path, class: "btn btn-fb btn-sns w-200px", method: :post do
+        = link_to user_facebook_omniauth_authorize_path, method: :post, class: "btn btn-fb btn-sns w-200px" do
           span.fab.fa-facebook-f.mr-2
           = "Facebookで#{msg}"
       .btn-sns
-        = link_to user_twitter_omniauth_authorize_path, class: "btn btn-tw btn-sns w-200px", method: :post do
+        = link_to user_twitter_omniauth_authorize_path, method: :post, class: "btn btn-tw btn-sns w-200px" do
           span.fab.fa-twitter.mr-2
           = "Twitterで#{msg}"
       .btn-sns
-        = link_to user_google_oauth2_omniauth_authorize_path, class: "btn btn-gg btn-sns w-200px", method: :post do
+        = link_to user_google_oauth2_omniauth_authorize_path, method: :post, class: "btn btn-gg btn-sns w-200px" do
           span.fab.fa-google.mr-2
           = "Googleで#{msg}"
