+++ config/initializers/constants.rb
@@ -1,5 +1,5 @@
 MAIN_TITLE = "GUNPLA-Type".freeze
 VALID_EMAIL_REGEX = /\A[\w+\-.]+@[a-z\d\-]+(\.[a-z\d\-]+)*\.[a-z]+\z/i.freeze
 VALID_PASSWORD_REGEX = /\A^[0-9_a-zA-Z-]*$\Z/.freeze
-SUGGEST_COUNT = 5
+AUTOCOMPETE_COUNT = 5
 HISTORIES_STOCK_LIMIT = 4
