+++ config/initializers/draper.rb
@@ -1,0 +1,1 @@
+Draper::CollectionDecorator.delegate :current_page, :total_pages, :limit_value, :total_count, :offset_value
